add confirmation for removing device and category

+++ src/app/category/category.component.ts
@@ -5,36 +5,31 @@ import { Router } from '@angular/router';
 @Component({
   selector: 'app-category',
   templateUrl: './category.component.html',
-  styleUrls: ['./category.component.scss']
+  styleUrls: ['./category.component.scss'],
 })
 export class CategoryComponent implements OnInit {
-
   categories: Category[] = [];
 
-  constructor(private service: CategoryService, private router: Router) { }
+  constructor(private service: CategoryService, private router: Router) {}
 
-  removeCategory(id: number){
+  removeCategory(id: number) {
     console.log('Categoria removida');
-
-    this.service.removeCategory(id).subscribe(
-      (result) =>{
-
-        // Retorna para a lista de categorias
-       this.router.navigateByUrl('categories');
-       // Refresh no componente
-       this.ngOnInit();
-
-      },
-      (error) => console.error(error)
-    );
+    if (confirm('Delete category?')) {
+      this.service.removeCategory(id).subscribe(
+        (result) => {
+          // Retorna para a lista de categorias
+          this.router.navigateByUrl('categories');
+          // Refresh no componente
+          this.ngOnInit();
+        },
+        (error) => console.error(error)
+      );
+    }
   }
-
-
 
   ngOnInit(): void {
-    this.service.listCategories().subscribe((categories: Category[]) =>{
+    this.service.listCategories().subscribe((categories: Category[]) => {
       this.categories = categories;
-    })
+    });
   }
-
 }
